refactor(domain): split learning payload mapping support
